+++ nameko_kafka/consumers/__init__.py
@@ -1,0 +1,4 @@
+from .default import DefaultConsumer
+from .exactly_once import ExactlyOnceConsumer
+from .least_once import AtLeastOnceConsumer
+from .most_once import AtMostOnceConsumer