+++ CHANGELOG.md
@@ -3,3 +3,9 @@
 - Added entrypoint for consuming Kafka messages
 - Added dependency provider for publishing Kafka messages
 - Added code coverage for the extension
+
+# v0.2
+
+- Fixed identifier of managed spawned thread in entrypoint 
+- Added support different commit strategies
+- Added MongoDB offset storage for exactly once delivery semantics